refactor: image 관련 공통 로직 common 폴더 하위로 이동
