+++ backend/src/main/java/com/gachi/gacha/server/gacha/application/GachaImageService.java
@@ -5,7 +5,7 @@ import com.gachi.gacha.server.gacha.domain.Gacha;
 import com.gachi.gacha.server.gacha.domain.GachaImage;
 import com.gachi.gacha.server.gacha.domain.GachaImageJpaRepository;
 import com.gachi.gacha.server.gacha.domain.GachaJpaRepository;
-import com.gachi.gacha.server.store.infra.config.ImageUploader;
+import com.gachi.gacha.server.common.infra.config.ImageUploader;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
